feat(animations-task-2): Finish task

+++ src/app/employee/employee.component.ts
@@ -2,19 +2,36 @@ import { Component, OnInit } from '@angular/core';
 import { EmployeesService } from '../employees.service';
 import { ActivatedRoute, Router } from '@angular/router';
 import { FormBuilder } from '@angular/forms';
+import { animate, state, style, trigger, transition, keyframes } from '@angular/animations';
 
 @Component({
   selector: 'app-employee',
   templateUrl: './employee.component.html',
   styleUrls: ['./employee.component.scss'],
+  animations: [
+    trigger('modalTrigger', [
+      state('popout', style({
+        height: '40%',
+        opacity: 1,
+      })),
+      state('hidden', style({
+        height: '0%',
+        opacity: 0,
+      })),
+      transition('hidden <=> popout', [
+        animate('300ms')
+      ])
+    ])
+  ]
 })
 export class EmployeeComponent implements OnInit {
 
   id;
-  private employee;
+  private deleteConfirm = false;
   employee$;
   private updating = false;
   updateForm;
+  animated = false;
 
   constructor(
     private employeesService: EmployeesService,
@@ -45,9 +62,22 @@ export class EmployeeComponent implements OnInit {
   }
 
   deleteThisEmployee() {
-    this.employeesService.delete(this.id).subscribe();
-    window.alert('Employee removed');
-    this.router.navigate(['/employees']);
+    this.animated = true;
+    if (this.deleteConfirm) {
+      this.employeesService.delete(this.id).subscribe();
+      window.alert('Employee removed');
+      this.animated = false;
+      this.router.navigate(['/employees']);
+    }
+  }
+
+  confirmDelete() {
+    this.deleteConfirm = true;
+    this.deleteThisEmployee();
+  }
+
+  hide() {
+    this.animated = false;
   }
 
   ngOnInit() {
